+++ src/telemetry.rs
@@ -27,9 +27,22 @@ pub fn configure<T>(config: &SubscriberConfig<T>) -> Tracer {
         }
     };
 
+    let collector_url = telemetry.collector_url.as_str();
+    // OTLP before version 0.15 didn't append a /v1/traces suffix, but started doing so there.
+    // For backwards compatibility we strip it from configurations that do have it
+    let collector_url = collector_url
+        // In case of a trailing slash strip it
+        .strip_suffix('/')
+        .unwrap_or(collector_url)
+        .strip_suffix("/v1/traces")
+        .unwrap_or(collector_url);
+
+    // Backport https://github.com/open-telemetry/opentelemetry-rust/pull/1553
+    let collector_url = collector_url.strip_suffix('/').unwrap_or(collector_url);
+
     let otlp_exporter = opentelemetry_otlp::new_exporter()
         .http()
-        .with_endpoint(telemetry.collector_url.as_str());
+        .with_endpoint(collector_url.to_string());
 
     let resource = Resource::new(vec![
         KeyValue::new("environment", config.env.to_string()),