+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.9.0-rc.1] - 2024-03-11
+
+### Changed
+
+- Automatically append /v1/traces to the collector endpoint
+
+This is only done if a /v1/traces suffix isn't already there, meaning old configurations should continue to function
+
+---
+
 ## [0.9.0-rc.0] - 2024-03-04
 
 ### Changed
@@ -137,7 +147,9 @@ If you are using Jaeger to collect traces locally on your machine, you will need
 
 
 
-[Unreleased]: https://github.com/primait/prima_tracing.rs/compare/0.9.0...HEAD
+
+[Unreleased]: https://github.com/primait/prima_tracing.rs/compare/0.9.0-rc.1...HEAD
+[0.9.0-rc.1]: https://github.com/primait/prima_tracing.rs/compare/0.9.0-rc.0...0.9.0-rc.1
 [0.9.0]: https://github.com/primait/prima_tracing.rs/compare/0.8.1...0.9.0
 [0.8.1]: https://github.com/primait/prima_tracing.rs/compare/0.8.0...0.8.1
 [0.8.0]: https://github.com/primait/prima_tracing.rs/compare/0.7.2...0.8.0